[species] refactor tests and creates ::species spec
Currently, the tests rely on hard-coded, in-clj-file data structures to
to test behaviour of monstera.entities.species namespace. This creates a
lot of unneeded boilerplate. To fix this, creates a new function aimed
to instantiate a valid data structure from EDN fixture files.

Also, with sights on future and to test this new fixtures-based
behaviour, a ::species spec is created within that particular namespace.